rename `TapeTask` to `differentiable-Any` & rename `differentiable-float` to `differentiable-Float`& rename `differentiable-double` to `differentiable-Double` & rename `differentiable-indarray` to `differentiable-INDArray`& rename `TapeTaskFactory` to `tapefactories` and add `differentiable` & fix generate unidoc error & remove `Closeables`
