+++ unidoc/src/main/scala/com/thoughtworks/Aux.scala
@@ -1,0 +1,6 @@
+package com.thoughtworks
+
+/**
+  * Make unidoc complete full path
+  */
+private class Aux
